Refactor: split host v2 attributes to submodules
commit-id:661c850f
