+++ scarb/src/compiler/plugin/proc_macro/v2/host/attribute/mod.rs
@@ -1,0 +1,6 @@
+mod inner_attribute;
+mod item_attribute;
+mod parse_attributes;
+
+pub use inner_attribute::*;
+pub use item_attribute::*;
